+++ studio/src/app/modals/editor/app-publish/app-publish.tsx
@@ -8,7 +8,7 @@ import {AuthUser} from '../../../models/auth-user';
 import {Deck} from '../../../models/deck';
 
 import {AuthService} from '../../../services/api/auth/auth.service';
-import {DeckEditorService} from '../../../services/api/deck/deck-editor.service';
+import {DeckEditorService} from '../../../services/editor/deck/deck-editor.service';
 import {DeckService} from '../../../services/api/deck/deck.service';
 import {ErrorService} from '../../../services/core/error/error.service';